requirements.txt
FreeSimpleGUI==5.2.0.post1
numpy==2.2.5
pillow==11.2.1
pyasn1==0.6.1
PySimpleGUI==5.0.10
rsa==4.9.1
